+++ SongWiki/ViewModel/LoginViewModel.swift
@@ -25,5 +25,3 @@ struct LoginViewModel {
         return imputNotEmpty ? .systemPurple : .systemPurple.withAlphaComponent(0.5)
     }
 }
-// agora vai
-// agora vai 2
